requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bertopic==0.16.4
bitsandbytes==0.45.2
bleach==6.2.0
blis==1.2.0
bs4==0.0.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
fake-useragent==2.0.3
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.17.0
fonttools==4.55.8
fqdn==1.5.1
fsspec==2024.12.0
googlesearch-python==1.3.0
googletrans==4.0.0rc1
h11==0.14.0
h2==3.2.0
hdbscan==0.8.40
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaleido==0.2.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
lxml==5.3.1
lxml_html_clean==0.4.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
more-itertools==10.6.0
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.24.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
sentence-transformers==3.4.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacey==0.1.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
Telethon==1.38.1
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.2
trio==0.29.0
trio-websocket==0.12.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
uri-template==1.3.0
urllib3==1.26.20
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
